+++ scripts/utils.py
@@ -23,35 +23,31 @@ def despine(ax, sides=['left', 'right', 'top', 'bottom']):
     [ax.spines[side].set_visible(False) for side in sides]
 
 
-def plot_africa_totals(data, colors=['blue', 'orangered', 'lawngreen']):
+def plot_africa_totals(data,
+                       colors=dict(Confirmed='blue', Deaths='orangered',
+                                   Recovered='lawngreen', Active='grey')):
     """Create lineplots of coronavirus case totals in Africa.
 
     Parameters:
     ----------
     data: pd.DateFrame
-        A dataframe of historic coronavirus cases in Africa.
-    colors: str, valid matplotlib color input
-        Colors to apply to the 'confirmed', 'deaths', and 'recovered'
+        A dataframe of cummulative totals of coronavirus case data in Africa,
+        indexed by date.
+    colors: dict, valid matplotlib color input
+        Colors to apply to the 'confirmed', 'deaths', 'recovered' and 'active'
         lineplots, respectively'.
     """
-    # Convert Date column values to Timestamps
-    data['Date'] = data['Date'].astype('datetime64')
+    latest_date = data.index[-1].strftime('%b %d, %Y')  # used in title
 
-    # Calculate totals for each day. After this, only the columns 'Confirmed'
-    # 'Deaths' and 'Recovered' remain, with 'Date' becoming the index.
-    data = data.groupby('Date').sum()
-
-    latest_date = data.index.max().strftime('%h %d, %Y')  # used in title
-
+    data['Active'] = data['Confirmed'] - data['Deaths'] - data['Recovered']
     # Create a matplotlib Figure, with lineplots for 'confirmed', 'recovered',
     # and 'deaths'
-    fig = Figure(figsize=(10, 5), dpi=200)
+    fig = Figure(figsize=(10, 5), dpi=180)
     fig.suptitle(f'Total Coronavirus Cases in Africa as at {latest_date}',
                  size=18)
     ax = fig.subplots()
-    for idx, label in enumerate(['Confirmed', 'Deaths', 'Recovered']):
-        ax.plot(data[label].index, data[label], lw=2, color=colors[idx],
-                label=label)
+    for label in data.columns:
+        ax.plot(data[label], lw=2, color=colors[label], label=label)
 
     # Annotate lineplots with current totals
     for line in ax.lines:
@@ -70,7 +66,7 @@ def plot_africa_totals(data, colors=['blue', 'orangered', 'lawngreen']):
 
     # Set the x-axis to show dates in 2-week intervals
     ax.xaxis.set_major_locator(mdates.WeekdayLocator(interval=2))
-    ax.xaxis.set_major_formatter(mdates.DateFormatter('%h %d'))  # e.g. Jan 20
+    ax.xaxis.set_major_formatter(mdates.DateFormatter('%b %d'))  # e.g. Jan 20
 
     fig.legend(loc='right')
     fig.savefig('images/africa_totals.png', bbox_inches='tight',
@@ -83,17 +79,15 @@ def plot_daily_confirmed(daily_data):
     Parameters:
     ----------
     daily_data: pd.DataFrame
-        A dataframe of daily coronavirus cases, with columns for 'Date',
+        A dataframe of daily coronavirus case data, with columns for 'Date',
         'Country/Region', and 'Confirmed'.
     """
     latest_date = daily_data['Date'].astype('datetime64').max()\
-                                    .strftime('%h %d, %Y')
+                                    .strftime('%b %d, %Y')
+    data = daily_data['Confirmed'].sort_values()
 
-    # Get and sort confirmed cases in ascending order
-    data = daily_data.set_index('Country/Region')['Confirmed'].sort_values()
-
-    # Create a horizontal barplot with country-confirmed-cases as bars
-    fig = Figure(figsize=(6, 12))
+    # Create a horizontal barplot with a bar for each country
+    fig = Figure(figsize=(6, 12), dpi=150)
     ax = fig.subplots()
     ax.set_title(f'Confirmed Coronavirus Cases by Country as at {latest_date}',
                  size=18)
@@ -112,20 +106,67 @@ def plot_daily_confirmed(daily_data):
 
 
 def plot_geoscatter(geo_data):
-    """Create a bubble map of coronavirus cases in Africa.
+    """Create a bubble map of active coronavirus cases in Africa.
 
     Parameters:
     ----------
     geodata: pd.DataFrame
-        A DataFrame with columns 'Lat' and 'Long', of latitude and longitude
-        values respectively, with corresponding 'Confirmed` case data.
+        A DataFrame with columns 'Lat' and 'Lon', of latitude and longitude
+        values respectively, having 'Active' case data.
     """
+    # On occassion, the Active cases column (= Confirmed - Recovered - Deaths)
+    # has negative values due to error(s). Negative values in the `size`
+    # arguement of `scatter_geo` raise a ValueError. The following ensures the
+    # values are non-negative:
+    geo_data['Active'] = geo_data['Active'].clip(lower=0)
+
     fig = px.scatter_geo(
-        geo_data, lat='Lat', lon='Long', scope='africa', size='Confirmed',
-        color='Confirmed', height=600, width=600, size_max=50,
+        geo_data, lat='Lat', lon='Lon', scope='africa', size='Active',
+        color='Active', height=600, width=600, size_max=50,
         color_continuous_scale=['cyan', 'yellow', 'orangered'],
-        title='Geographic Scatter-plot of Confirmed Cases'
+        title='Geographic Scatter-plot of Active Coronavirus Cases'
     )
     # Save geographical scatter-plot as png image file
     with open('images/geo_scatter.png', 'wb') as pic:
         pic.write(fig.to_image(format='png', scale=2))
+
+
+def plot_daily_stats(daily_stats):
+    """Plot a pair of horizontal bar-plots for coronavirus incidence-rate &
+    case-fatality-ratio in Africa.
+
+    Parameters:
+    ----------
+    daily_stats: pd.DataFrame
+        A DataFrame with the incidence-rate and case-fatality-ratio info,
+        indexed by country.
+    """
+    fig = Figure(figsize=(12, 12), tight_layout=True)
+    ax1, ax2 = fig.subplots(nrows=1, ncols=2)
+
+    # Horizontal barplot of incidence rate
+    daily_stats['Incidence Rate'].sort_values()\
+                                 .plot.barh(ax=ax1, color='darkviolet')
+    ax1.set_title('Incidence Rate (Cases Per 100,000 Persons)', size=16)
+    ax1.set_ylabel('Country/Region')
+    despine(ax1)  # remove bordering box
+    ax1.get_xaxis().set_visible(False)  # hide x-axis
+
+    # Horizontal barplot of case-fatality-ratio
+    daily_stats['Case - Fatality Ratio'].sort_values()\
+                                        .plot.barh(ax=ax2, color='dimgray')
+    ax2.set_title(
+        'Case-Fatality Ratio \n(№ Recorded Deaths / № Cases * 100%)',
+        size=16)
+    ax2.set_ylabel('Country/Region')
+    despine(ax2)  # remove bordering box
+    ax2.get_xaxis().set_visible(False)  # hide x-axis
+
+    # Annotate the bars
+    for p in ax1.patches:
+        ax1.annotate(f'{p.get_width():,.2f}', (p.get_width(), p.get_y()))
+
+    for p in ax2.patches:
+        ax2.annotate(f'{p.get_width():.2f}%', (p.get_width(), p.get_y()))
+
+    fig.savefig('images/stats.png', transparent=True)